refactor(bench): extract snapshot module from bench_serving
Final slice of the split. Move the snapshot + git-baseline-compare subsystem
into snapshot.rs: the SNAPSHOT_*/REGRESSION_* consts, run_snapshot,
render_snapshot_text, run_compare, render_comparison, and the
git/gpu/date/delta provenance helpers. snapshot glob-imports its siblings.

main.rs is now a ~330-line orchestrator (command_seed, kimi_parallel_config,
dispatch, main, the dsv2-lite test module). Trimmed its now-unused imports;
qualified the lone kimi-gated ensure! call site so the import could go.

Pure code motion, no behavior change. Verified: default build clean and all
11 deepseek-v2-lite unit tests pass (only the pre-existing ParallelConfig
kimi-gated unused-import warning remains).
